Clarify Gateway Translation, xDS  and e2e tests (#6504)
* Part 1: introduce predicate, improve some comments around translation

* more comments around translation

* add comments around testing

* small cleanup

* function cleanup

* cleanup mock function

* pass writeNamesapce properly

* pass writeNamesapce properly

* gofmt

* update readme

* Apply suggestions from code review


* Update projects/gloo/pkg/xds/README.md


* codegen
