+++ projects/gateway/pkg/reconciler/proxy_reconciler_test.go
@@ -64,8 +64,10 @@ var _ = Describe("ReconcileGatewayProxies", func() {
 	)
 
 	genProxy := func() {
-		tx := translator.NewDefaultTranslator(translator.Opts{})
-		proxy, reports = tx.Translate(context.TODO(), "proxy-name", ns, snap, snap.Gateways)
+		tx := translator.NewDefaultTranslator(translator.Opts{
+			WriteNamespace: ns,
+		})
+		proxy, reports = tx.Translate(context.TODO(), "proxy-name", snap, snap.Gateways)
 
 		proxyToWrite = GeneratedProxies{proxy: reports}
 	}
