+++ projects/gateway/pkg/reporting/check_source_reports_test.go
@@ -26,8 +26,10 @@ var _ = Describe("CheckSourceReports", func() {
 	)
 	BeforeEach(func() {
 		snap = samples.SimpleGatewaySnapshot(&core.ResourceRef{Name: ignored, Namespace: ignored}, ignored)
-		tx := translator.NewDefaultTranslator(translator.Opts{})
-		proxy, reports = tx.Translate(context.TODO(), ignored, ignored, snap, snap.Gateways)
+		tx := translator.NewDefaultTranslator(translator.Opts{
+			WriteNamespace: ignored,
+		})
+		proxy, reports = tx.Translate(context.TODO(), ignored, snap, snap.Gateways)
 	})
 	It("returns true when all the sources for the config object are error-free", func() {
 		accepted, err := AllSourcesAccepted(reports, proxy.Listeners[0])
