+++ projects/gateway/pkg/syncer/translator_syncer_test.go
@@ -386,7 +386,7 @@ var _ = Describe("TranslatorSyncer", func() {
 
 			ctx = settingsutil.WithSettings(ctx, settings)
 
-			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", "gloo-system", snap, gomock.Any()).
+			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", snap, gomock.Any()).
 				Return(proxy, nil)
 
 			ts.GeneratedDesiredProxies(ctx, snap)
@@ -395,7 +395,7 @@ var _ = Describe("TranslatorSyncer", func() {
 		})
 
 		It("should not compress proxy spec when setttings are not set", func() {
-			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", "gloo-system", snap, gomock.Any()).
+			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", snap, gomock.Any()).
 				Return(proxy, nil)
 
 			ts.GeneratedDesiredProxies(ctx, snap)
@@ -404,7 +404,7 @@ var _ = Describe("TranslatorSyncer", func() {
 		})
 		It("should truncate proxy status when limit is set", func() {
 
-			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", "gloo-system", snap, gomock.Any()).
+			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", snap, gomock.Any()).
 				Return(proxy, nil)
 			ts.proxyStatusMaxSize = "5"
 			ts.GeneratedDesiredProxies(ctx, snap)
@@ -412,7 +412,7 @@ var _ = Describe("TranslatorSyncer", func() {
 		})
 		It("should not truncate proxy status when limit is not set", func() {
 
-			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", "gloo-system", snap, gomock.Any()).
+			mockTranslator.EXPECT().Translate(gomock.Any(), "gateway-proxy", snap, gomock.Any()).
 				Return(proxy, nil)
 			ts.GeneratedDesiredProxies(ctx, snap)
 			Expect(proxy.Metadata.Annotations).NotTo(HaveKey(compress.ShortenKey))
