+++ projects/gateway/pkg/translator/mocks/mock_translator.go
@@ -38,16 +38,16 @@ func (m *MockTranslator) EXPECT() *MockTranslatorMockRecorder {
 }
 
 // Translate mocks base method.
-func (m *MockTranslator) Translate(arg0 context.Context, arg1, arg2 string, arg3 *v1.ApiSnapshot, arg4 v1.GatewayList) (*v10.Proxy, reporter.ResourceReports) {
+func (m *MockTranslator) Translate(arg0 context.Context, arg1 string, arg2 *v1.ApiSnapshot, arg3 v1.GatewayList) (*v10.Proxy, reporter.ResourceReports) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Translate", arg0, arg1, arg2, arg3, arg4)
+	ret := m.ctrl.Call(m, "Translate", arg0, arg1, arg2, arg3)
 	ret0, _ := ret[0].(*v10.Proxy)
 	ret1, _ := ret[1].(reporter.ResourceReports)
 	return ret0, ret1
 }
 
 // Translate indicates an expected call of Translate.
-func (mr *MockTranslatorMockRecorder) Translate(arg0, arg1, arg2, arg3, arg4 interface{}) *gomock.Call {
+func (mr *MockTranslatorMockRecorder) Translate(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Translate", reflect.TypeOf((*MockTranslator)(nil).Translate), arg0, arg1, arg2, arg3, arg4)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Translate", reflect.TypeOf((*MockTranslator)(nil).Translate), arg0, arg1, arg2, arg3)
 }
