+++ projects/gateway/pkg/validation/validator_test.go
@@ -36,9 +36,11 @@ var _ = Describe("Validator", func() {
 	)
 
 	BeforeEach(func() {
-		t = translator.NewDefaultTranslator(translator.Opts{})
 		ns = "my-namespace"
-		v = NewValidator(NewValidatorConfig(t, vf, ns, false, false))
+		t = translator.NewDefaultTranslator(translator.Opts{
+			WriteNamespace: ns,
+		})
+		v = NewValidator(NewValidatorConfig(t, vf, false, false))
 		mValidConfig = utils.MakeGauge("validation.gateway.solo.io/valid_config", "A boolean indicating whether gloo config is valid")
 	})
 
@@ -334,7 +336,7 @@ var _ = Describe("Validator", func() {
 
 			Context("allowWarnings=false", func() {
 				BeforeEach(func() {
-					v = NewValidator(NewValidatorConfig(t, acceptProxy, ns, true, false))
+					v = NewValidator(NewValidatorConfig(t, acceptProxy, true, false))
 				})
 				It("rejects a vs with missing route table ref", func() {
 					v.validationFunc = warnProxy
