+++ projects/gloo/pkg/syncer/setup/setup_syncer.go
@@ -743,7 +743,6 @@ func RunGlooWithExtensions(opts bootstrap.Opts, extensions Extensions, apiEmitte
 	gwValidationSyncer := gwvalidation.NewValidator(gwvalidation.NewValidatorConfig(
 		gatewayTranslator,
 		validator.Validate,
-		gwOpts.WriteNamespace,
 		ignoreProxyValidationFailure,
 		allowWarnings,
 	))
