+++ projects/gloo/pkg/xds/README.md
@@ -8,7 +8,27 @@
 
 Gloo Edge is an xDS server. It maintains a snapshot-based, in-memory cache and responds to xDS requests with the resources that are requested.
 
-The following discovery services are supported by Gloo Edge:
+### Snapshot
+
+A [snapshot](https://github.com/solo-io/solo-kit/blob/97bd7c2c67420a6d99bb96f220f2e1a04c6d8a0d/pkg/api/v1/control-plane/cache/snapshot.go#L43) is a versioned group of resources. In Gloo Edge, we rely on an [Envoy snapshot](https://github.com/solo-io/gloo/blob/1f457f4ef5f32aedabc58ef164aeea92acbf481e/projects/gloo/pkg/xds/envoy_snapshot.go#L39), a snapshot of the xDS resources that Gloo serves to Envoy.
+
+### Snapshot Cache
+
+A [snapshot cache](https://github.com/solo-io/solo-kit/blob/97bd7c2c67420a6d99bb96f220f2e1a04c6d8a0d/pkg/api/v1/control-plane/cache/simple.go#L70) maintains a single versioned snapshot per key. It also responds to open xDS requests.
+
+### xDS Callbacks
+
+[xDS callbacks](https://github.com/solo-io/solo-kit/blob/97bd7c2c67420a6d99bb96f220f2e1a04c6d8a0d/pkg/api/v1/control-plane/server/generic_server.go#L76) are a set of callbacks that are invoked asynchronously during the lifecycle of an xDS request.
+
+Gloo Edge open source does not define any xDS callbacks. However, these callbacks are a type of [extension that can be injected at runtime](https://github.com/solo-io/gloo/blob/75c0ee0f3b70258d0013364e82489f570685e1d7/projects/gloo/pkg/syncer/setup/setup_syncer.go#L393). Gloo Edge Enterprise defines xDS callbacks, and injects them into the Control Plane at runtime.
+
+### Server
+
+An [xDS server](https://github.com/solo-io/solo-kit/blob/97bd7c2c67420a6d99bb96f220f2e1a04c6d8a0d/pkg/api/v1/control-plane/server/generic_server.go#L52) defines a set of handlers for streaming discovery requests.
+
+## xDS Services
+
+The xDS server is configured to expose the following discovery services in Gloo Edge:
 
 ### ListenerDiscoveryService
 
@@ -48,12 +68,6 @@ To guarantee that proxies initiate requests for the snapshot they want, it is cr
 
 Proxies identify the cache key that they are interested in by specifying their `node.metadata.role` to the cache key using the above naming pattern. An example of this can be found in the [bootstrap configuration for proxies](https://github.com/solo-io/gloo/blob/0eec04dc0486976fc89bac314b0fd9eccd5261f5/install/helm/gloo/templates/9-gateway-proxy-configmap.yaml#L45)
 
-## xDS Callbacks
-
-[xDS callbacks](https://github.com/solo-io/solo-kit/blob/97bd7c2c67420a6d99bb96f220f2e1a04c6d8a0d/pkg/api/v1/control-plane/server/generic_server.go#L76) are a set of callbacks that are invoked asynchronously during the lifecycle of an xDS request.
-
-Gloo Edge open source does not define any xDS callbacks. However, these callbacks are a type of [extension that can be injected at runtime](https://github.com/solo-io/gloo/blob/75c0ee0f3b70258d0013364e82489f570685e1d7/projects/gloo/pkg/syncer/setup/setup_syncer.go#L393). Gloo Edge Enterprise defines xDS callbacks, and injects them into the Control Plane at runtime.
-
 ## xDS Debugging
 
 Debugging xDS behavior can be challenging, below are a few techniques to help:
@@ -69,4 +83,5 @@ Gloo supports running in [dev mode](https://github.com/solo-io/gloo/blob/6994b41
 ## Useful information
 
 - [Hoot YouTube series about xDS](https://www.youtube.com/watch?v=S5Fm1Yhomc4)
+- [Envoy xDS](https://www.envoyproxy.io/docs/envoy/latest/api-docs/xds_protocol)
 - [Hoot Repository](https://github.com/solo-io/hoot)
